refactor: rename recorder reference video component
